Add `Index` Post curl script
Curl script is working in local environment.
This action returns an array of Posts objects with attribute editable
with a value of `false` in the case of Posts owned by other users and
`true` in the case of Posts owned by the current user.

Note: When providing the ID of a post not owned by the current user, the
actual error returned by the `Update` curl script is `Not Found`. The
`Update` curl script on the previous commit was mistakenly tested
without providing the current user's token.